Labs8-Randomizer/requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.9
django-allauth==0.38.0
django-cors-headers==2.4.0
django-currentuser==0.2.3
django-heroku==0.3.1
django-rest-api==0.1.5
django-rest-auth==0.9.3
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
oauthlib==2.1.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.1
whitenoise==4.1.1
